# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de la API",
        "description": "Documentación de la API proyecto CDMI",
        "contact": {
            "email": "tucorreo@ejemplo.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión de usuario",
                "description": "Inicia sesión y devuelve un token de autenticación",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "correo",
                                    "contraseña",
                                    "rol"
                                ],
                                "properties": {
                                    "correo": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "usuario@example.com"
                                    },
                                    "contraseña": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "rol": {
                                        "type": "string",
                                        "enum": [
                                            "administrador",
                                            "usuario"
                                        ],
                                        "example": "usuario"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso"
                    },
                    "401": {
                        "description": "Correo no registrado o contraseña incorrecta"
                    },
                    "403": {
                        "description": "Rol incorrecto"
                    },
                    "422": {
                        "description": "Error de validación de datos"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/api/verify": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Verificar sesión del usuario",
                "description": "Devuelve la información del usuario autenticado",
                "operationId": "7b61aeb1287dc38252412e08a8c7202a",
                "responses": {
                    "200": {
                        "description": "Usuario autenticado encontrado"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión del usuario",
                "description": "Elimina el token de acceso del usuario autenticado",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Cierre de sesión exitoso"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ayuda": {
            "get": {
                "tags": [
                    "Ayuda"
                ],
                "summary": "Obtener todas las donaciones",
                "description": "Devuelve un listado de todas las donaciones registradas",
                "operationId": "cb062cc2bff062949fb632239735ae83",
                "responses": {
                    "200": {
                        "description": "Listado de donaciones obtenido exitosamente"
                    }
                }
            },
            "post": {
                "tags": [
                    "Ayuda"
                ],
                "summary": "Registrar una nueva donacion",
                "description": "Crea una nueva opción de donación",
                "operationId": "871181857750ca69605a9db59dca478a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "fecha",
                                    "monto",
                                    "causa"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Donación Escolar"
                                    },
                                    "fecha": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-04-26"
                                    },
                                    "monto": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "50000"
                                    },
                                    "causa": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Apoyo a niños indígenas"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Donacion creada exitosamente"
                    },
                    "422": {
                        "description": "Error en la validación de datos"
                    }
                }
            }
        },
        "/api/ayuda/{ayudaId}/registrar-usuario": {
            "post": {
                "tags": [
                    "Ayuda"
                ],
                "summary": "Registrar usuario en una donacion",
                "description": "Asocia al usuario autenticado con una opción de donacion específica",
                "operationId": "48e5c8deb0a50c7afb5dce1899669de5",
                "parameters": [
                    {
                        "name": "ayudaId",
                        "in": "path",
                        "description": "ID de la donacion a la cual el usuario quiere donar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Donación realizada con éxito"
                    },
                    "404": {
                        "description": "Opción de donacion no encontrada"
                    },
                    "409": {
                        "description": "El usuario ya ha donado a esta donacion"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}